created one api client with getPlayer and getStats
