+++ server/src/routes/api/v1/playersRouter.js
@@ -1,6 +1,6 @@
 import express from "express"
 
-import PlayersClient from "../../../apiClient/PlayersClient.js"
+import balldontlieClient from "../../../apiClient/balldontlieClient.js"
 import PlayerSerializer from "../../../serializers/PlayerSerializer.js"
 
 const playersRouter = new express.Router()
@@ -8,7 +8,7 @@ const playersRouter = new express.Router()
 playersRouter.get("/", (req, res) => {
   const playerName = req.query.playerName
 
-  PlayersClient.getPlayer(playerName)
+  balldontlieClient.getPlayer(playerName)
     .then((data) => {
       if (data.error) {
         console.log(`Error from balldontlie api: ${data.error}`)
